Fixed data loading problems
Data must be loaded on the index file
